Organize structure of tests
